pyproject.toml
[tool.poetry]
name = "initial-repo"
version = "1.0.0"
description = "DESRC_HERE"
authors = ["jay naik <jay@bdlabs.ai>"]

[tool.poetry.dependencies]
python = "3.9.5"
pathlib2 = "2.3.7"
numpy = "1.20.2"
pandas = "1.2.5"
python-dotenv = "0.17.0"
requests = "2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
